requirements.txt
anyio==3.6.1
blis==0.7.8
catalogue==2.0.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cymem==2.0.6
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
packaging==21.3
thinc==8.0.17
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
wasabi==0.9.1
watchfiles==0.15.0
websockets==10.3
gunicorn==20.0.4
Flask==1.0.2
fast-autocomplete==0.9.0
pylev==1.4.0
joblib==1.1.0
jsons==1.6.3
pymongo==4.2.0
python-dotenv==1.0.0
openai==0.26.5
pymongo[srv]